feat: Support for changing the color of the boxes in the heatmap from changing the postType.
